refractor: FieldReplace functions -> fieldReplacersUtil.ts
