+++ frontend/src/util/helper.ts
@@ -377,35 +377,6 @@ export const isJsonString = (str: string) => {
   return true;
 };
 
-
-
-export const populateIntroField = (input: string): string => {
-  return input
-}
-
-export const populateNameField = (input: string, studentname: string): string => {
-  return input
-    .replace('{studentname}', studentname)
-}
-
-export const populateFooterField = (input: string, classname: string, datebootcamp: string): string => {
-  return input
-    .replace('${classname}', classname)
-    .replace('${datebootcamp}', datebootcamp);
-}
-
-export const populateField = (input: string, classname: string, datebootcamp: string, studentname: string): string => {
-  return input
-    .replace('{classname}', classname)
-    .replace('{datebootcamp}', datebootcamp)
-    .replace('{studentname}', studentname)
-}
-
-export const populateIdField = (input: string, verificationCode: string): string => {
-  return input
-    .replace('{id}', verificationCode)
-}
-
 export const delay = (ms: number) => new Promise(resolve => setTimeout(resolve, ms));
 
 export function mapBootcampToSaltData(bootcamp: BootcampResponse, template: TemplateResponse ): SaltData {
